pyproject.toml
[project]
name = "taisp-nerd-bot"
version = "0.1.0"
description = "Telegram bot for NTU TAISP students to track assignment deadlines"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot[job-queue]>=20.0",
]

[project.scripts]
bot = "bot:main"
